requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3